pyproject.toml
[project]
name = "ask-my-document-using-mongo-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "pymongo>=4.15.0",
    "pypdf>=6.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "uvicorn>=0.35.0",
]
